Refactor rgb_led
Renamed LEDLight to RgbLed.  Moved analogWrite() code to an external pollyfil source file.
